+++ Campaign/extensions/campaign-module/src/Providers/StatsServiceProvider.php
@@ -10,6 +10,8 @@ use Illuminate\Support\ServiceProvider;
 
 class StatsServiceProvider extends ServiceProvider
 {
+    private const CONNECT_TIMEOUT = 3;
+
     /**
      * Bootstrap the application services.
      *
@@ -30,8 +32,8 @@ class StatsServiceProvider extends ServiceProvider
         $this->app->bind(StatsContract::class, function (Application $app) {
             $client = new Client([
                 'base_uri' => config('services.remp.beam.segments_addr'),
-                'timeout' => config('services.remp.beam.segments_timeout'),
-                'connect_timeout' => 1,
+                'timeout' => config('services.remp.beam.segments_timeout') + self::CONNECT_TIMEOUT,
+                'connect_timeout' => self::CONNECT_TIMEOUT,
             ]);
 
             return new Stats($client);